+++ sickbeard/server/web/manage/__init__.py
@@ -1,0 +1,3 @@
+
+from sickbeard.server.web.manage.base import Manage
+from sickbeard.server.web.manage.searches import ManageSearches
